% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastlm.R
\name{fastlm_chol}
\alias{fastlm_chol}
\title{Fast Linear Regression using Cholesky Decomposition}
\usage{
fastlm_chol(X, y, intercept = TRUE)
}
\arguments{
\item{X}{Numeric matrix of predictors.}

\item{y}{Numeric vector of response values.}

\item{intercept}{Logical; whether to include an intercept (default = TRUE).}
}
\description{
Fits a linear regression model using Cholesky decomposition
(usually faster than QR).
}
\examples{
data(mtcars)
X <- as.matrix(mtcars[, c("wt", "hp")])
y <- mtcars$mpg
fit <- fastlm_chol(X, y)
fit$coefficients
}
