% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastlm.R
\name{fastlm_qr}
\alias{fastlm_qr}
\title{Fast Linear Regression using QR Decomposition}
\usage{
fastlm_qr(X, y, intercept = TRUE)
}
\arguments{
\item{X}{Numeric matrix of predictors.}

\item{y}{Numeric vector of response values.}

\item{intercept}{Logical; whether to include an intercept (default = TRUE).}
}
\value{
A list of class \code{"fastlm"} containing:
\item{coefficients}{Estimated regression coefficients}
\item{fitted.values}{Fitted (predicted) values}
\item{residuals}{Residuals}
\item{sigma}{Estimated residual standard error}
\item{df.residual}{Residual degrees of freedom}
\item{r.squared}{R-squared}
\item{adj.r.squared}{Adjusted R-squared}
\item{vcov}{Variance-covariance matrix of coefficients}
\item{se}{Standard errors of coefficients}
}
\description{
Fits a linear regression model using QR decomposition.
}
\examples{
data(mtcars)
X <- as.matrix(mtcars[, c("wt", "hp")])
y <- mtcars$mpg
fit <- fastlm_qr(X, y)
fit$coefficients
}
