% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastlm.R
\name{print.fastlm}
\alias{print.fastlm}
\title{Print method for fastlm objects}
\usage{
\method{print}{fastlm}(x, ...)
}
\arguments{
\item{x}{A \code{fastlm} model.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for fastlm objects
}
