% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastlm.R
\name{print.summary.fastlm}
\alias{print.summary.fastlm}
\title{Print method for summary.fastlm objects}
\usage{
\method{print}{summary.fastlm}(x, ...)
}
\arguments{
\item{x}{A summary of \code{fastlm} model.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for summary.fastlm objects
}
