% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastlm.R
\name{summary.fastlm}
\alias{summary.fastlm}
\title{Summary method for fastlm objects}
\usage{
\method{summary}{fastlm}(object, ...)
}
\arguments{
\item{object}{A \code{fastlm} model.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input \code{fastlm} object invisibly.
}
\description{
Summary method for fastlm objects
}
